requirements.txt
selenium
pandas
pyautogui
pystray
Pillow
python-dotenv
schedule
customtkinter
